persistence: extract process loop
